requirements.txt
torch~=1.8.1
tqdm~=4.59.0
scikit-learn~=0.24.2
pandas~=1.2.4
pillow~=8.2.0
faiss~=1.6.3
tensorflow~=2.3.2
katna~=0.9.2
ffmpeg-python==0.2.0
